correcting the inference framework for comparisons
In this commit, it's corrected the random labelling approach for either spatial and data approach. A new inference function was developed called 'compare segregation' and can be used to compare either two different spatial contexts or two points in time for the same region. The code still needs to be clean.