adding core_data public argument
this commit adds a core_data public argument in all functions and adapt everything that rely on each function. For example, if a measure relies on the _dissim function, it will access the value of the dissimilarity with [0] index. In addition, in the definition of the class, I put an auxiliary object (called 'aux') in order to avoid multiple calls of the same function in order to be more efficient.