adds a wrapper function to perform inference between two segregation measures
this commit adds the main inference function to perform inference for two segregation measures. The null_approach can be either random_data or random_geometry which affects differently the iterations. This functions excpects the user to put as input two segregation previously already built in this module. In addition, since some measures rely on different parameters (for example, alpha, beta, b, m, etc.) the user need to put explicitly again the arguments in this function with **kwargs.